+++ packages/ragbits-chat/src/ragbits/chat/interface/_interface.py
@@ -150,6 +150,17 @@ class ChatInterface(ABC):
         expected_signature = ChatInterface._sign_state(state)
         return hmac.compare_digest(expected_signature, signature)
 
+    async def setup(self) -> None:  # noqa: B027
+        """
+        Setup the chat interface.
+
+        This method is called after the chat interface is initialized and before the chat method is called.
+        It is used to setup the chat interface, such as loading the model or initializing the vector store.
+
+        This method is optional and can be overridden by subclasses.
+        """
+        pass
+
     @abstractmethod
     async def chat(
         self,