+++ packages/ragbits-chat/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add setup method to ChatInterface (#586)
+
 ## 0.19.1 (2025-05-27)
 
 ### Changed